Fiw more msvc warnings

+++ src/core/db.cc
@@ -104,7 +104,7 @@ RowValue::RowValue( pg_result* res, size_t nrow )
 Value RowValue::operator [] ( size_t fn ) {
     BOOST_ASSERT( fn < ( size_t )PQnfields( res_ ) );
     return Value( PQgetvalue( res_, nrow_, fn ),
-                  PQgetlength( res_, nrow_, fn ),
+                  ( size_t )PQgetlength( res_, nrow_, fn ),
                   PQgetisnull( res_, nrow_, fn ) != 0
                   );
 }
